Move exceptions into exception.core package.
One step closer to be Java 9 compatible.
Fixes #416
