+++ core/src/main/java/org/neo4j/ogm/exception/core/InvalidDepthException.java
@@ -11,21 +11,17 @@
  *  conditions of the subcomponent's license, as noted in the LICENSE file.
  */
 
-package org.neo4j.ogm.exception;
+package org.neo4j.ogm.exception.core;
 
 /**
- * Unrecoverable exception thrown when there is a user error in entity metadata (conflicting annotations,
- * missing id fields etc.)
- *
- * @author Frantisek Hartman
+ * @author Luanne Misquitta
  */
-public class MetadataException extends RuntimeException {
+public class InvalidDepthException extends RuntimeException {
 
-    public MetadataException(String message) {
-        super(message);
+    public InvalidDepthException() {
     }
 
-    public MetadataException(String message, Throwable cause) {
-        super(message, cause);
+    public InvalidDepthException(String message) {
+        super(message);
     }
 }
