+++ core/src/main/java/org/neo4j/ogm/metadata/reflect/EntityFactory.java
@@ -17,8 +17,8 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 
-import org.neo4j.ogm.exception.BaseClassNotFoundException;
-import org.neo4j.ogm.exception.MappingException;
+import org.neo4j.ogm.exception.core.BaseClassNotFoundException;
+import org.neo4j.ogm.exception.core.MappingException;
 import org.neo4j.ogm.metadata.ClassInfo;
 import org.neo4j.ogm.metadata.MetaData;
 import org.neo4j.ogm.model.Edge;
