+++ core/src/main/java/org/neo4j/ogm/session/request/NodeQueryBuilder.java
@@ -21,7 +21,7 @@ import java.util.Map;
 
 import org.neo4j.ogm.cypher.BooleanOperator;
 import org.neo4j.ogm.cypher.Filter;
-import org.neo4j.ogm.exception.MissingOperatorException;
+import org.neo4j.ogm.exception.core.MissingOperatorException;
 
 /**
  * @author Jasper Blues
