+++ test/src/main/java/org/neo4j/ogm/testutil/MultiDriverTestClass.java
@@ -25,7 +25,7 @@ import org.neo4j.ogm.driver.Driver;
 import org.neo4j.ogm.drivers.bolt.driver.BoltDriver;
 import org.neo4j.ogm.drivers.embedded.driver.EmbeddedDriver;
 import org.neo4j.ogm.drivers.http.driver.HttpDriver;
-import org.neo4j.ogm.exception.ConfigurationException;
+import org.neo4j.ogm.exception.core.ConfigurationException;
 import org.neo4j.ogm.session.SessionFactory;
